seperated maps into the data
removed 'selection' flag from retrieve_from_scene.  Now forcing a
selection.  Though I am using zQuery to find solver.  Default behavior
is to return a solver if nothing is selected so it works well.

New map class which deals with maps and interpolating.  This gets put in
the data of nodeCollection.

some cleanup and more logging