removed 'selection' argument from retrieve_from_scene_selection
This forces a selection in maya in order to retrieve stuff.  Reason
for this is because this is much safer in terms of name collisions.
With a selection argument the user could pass a short name and have an
unresolvable name collision.  All retrieve methods are storing long and
short names