zFiber creation works no longer based on existing name
zFiber creation much more robust.  If it finds existing fibers on nodes it uses all those
instead of creating new ones and stores the mObject.  This way it wont get
tripped up on maya naming it something other then expected as it uses mObjects
for setting attributes and applying maps.