Removed 'type' properties.
going with a public attr as 'type'  Previous I had TYPE and _type.
type is a class attr.  This means it is not part of __dict__.
I need it a class attr AND I need to save it.  I need it a class attr
so I can use the factories to search for the 'type' without
having to instantiate it.  I need to save it because when I look
through json data I need to know type to pass to factory.

So, in itit I am doing a
self.type = self.type

Which seems to push the class attr to __dict__.  Confusing.