stripped out checking map validity out of zBuilder.
put the core logix in zMaya and zBuilder calls it.  Also added it
to zBuilder.util so can be run outside of zBuilder.

Used to raise an error if it found a problem.  This was not helpful.
So now it turns off offending tissues and returns bad maps.  Can be
run on any ziva scene with a util.check_map_validity()