changed class names for clarity and added a dgnode class
base parameter has been changed to make it much simpler.  Now it
just deals with serialization and names.  dgnode is new and is the
new 'base' which inherits from base.  So map and mesh are not a
dgnode and dont inherit from that, they inherit from new base.

Changed classn name of BaseParameter to Base
DeformerBaseParameter to Deformer