changed get_maps and get_meshes to get_parameters
When retrieving turing this off wont get mesh or map info in this case
Reason you may want to do this.  Much much faster if all you want to
do is store attributes.  Good to build a ui off this as you wont need maps or meshes
Though the refernces are still there.