BugFix: materials are in wrong order when using retrieve_from_scene
The nodes must be stored in the 'correct' order so that when they
are applied to the scene, they come out in the right order.
It's the job of retrieve_XXX to make sure they are in the correct order.
retrieve_from_scene() is calling this zQuery method to get everything
at once, but they come out in the wrong order. They need to be in
'forward' order, but listHistory returns things in 'backwards' order
(depth-first search backwards from the start node).

The fix is easy enough: reverse the result.