Refactoring code for the serialization and deserialization
Previously, different places in code had to be aware of contents of
other nodes.  Specifically this was happening with anything that
inherited from dg_node which dealt with the mobjects.  For this to
work I am overwriting .serialize and .deserialize in dg_node so
now it is the only node that knows anything about mobjects.