VFXACT-395 Rename "Cache" to "Simulation RAM Cache" in menu
Because people keep thinking this is a good way to save results.
It's not. Maybe they think this because we've named it "Cache", 
which is easy to interpret in lots of different ways. 
So, let's rename it to be a lot more specific about what it's doing.
It's a simulation cache, not a result cache or a point cache.
Also, it's a RAM cache. It doesn't automatically write to/from disk without scripts.