VFXACT-409 Changing comparison for first body item
The bodies in zBuilder are all built at once in one ziva command.  This is to speed up the build process as it is faster to do that then to build each one individually.  In order to do that we only built the bodies IF it is first scene item of that type in the bundle.  It was doing a == check but now an IS check is more reliable.