VFXACT-409 Removing an extra loop when applying attrs to zTissue
The applying of attributes was happening in an extra check if its first item.  Moved to same check as 'build_multiple'.  This happens here because it only needs to run once, AND it needs to run regardless if there are zTissues to build.  This case comes up when you copy/paste for example.  Any time you want to apply settings when the tissue is in scene already.