Conforming attributes menu to maps menu
As of right now we have the right click menu working on single selection.  We want to enable multiple selection.  The commit makes the copy paste attribute actions similar to copy and paste maps.  Previously the attributes were looping through the items for each method, the loop for multiple selection is better further upstream so we only need to do loop once as opposed to everything.