VFXACT-461 Replace 'StandardError' with 'Exception'
StandardError was removed in Python 3, we should use Exception instead.
(See: http://python3porting.com/differences.html#standarderror)

This code is Python 2 code, so it works fine. 
However, we'd like it to also be valid Python 3 code for two reasons:
1) So we can smoothly keep using it when Maya switches to Python 3.
2) We'd like to interpret it with Python 3 during document generation,
so we can use the same version of Sphinx for all of our Python code doc gen.