VFXACT-461 Use generated modules.rst instead of zBuilder.rst
modules.rst is the top-level table of contents for the auto-generated docs.
I _think_ that's what we're supposed to be using, not zBuilder.rst.

I'm doing this now because we're seeing a warning about modules.rst being unused.
This is one way to fix it.
The other way is to tell sphinx not to generate modules.rst, but that seems to go
against the way it's designed. So, I went this way instead.