VFXACT-490 Fixing a bug for this ticket.
Previously to this commit we were checking if the zRestNode existed by checking scene for the name of it.  Now it is checking for existing zRestNodes on the specific mesh.  

What was happening is during a string_replace sometimes the zRestShape node name would not change SO it would be in scene but not associated with proper (name changed in this case) mesh so nothing would happen.  Now it ignores name and checks mesh.