VFXACT-535 remove a list comprehension
We were treating parameter_args as a list where each item we search for.  Fact is only first element is name so could simplify a bit.  The logic is in a try: except: with a raised ValueError meaning that the item is not in list.  In that case we create a node.