VFXACT-530 Filtering out attributes that are not writable
If an attribute is not writable we are now filtering it out during the retrieving.  If it is not writable there is nothing we can do about it so no sense storing it.  If stored we would need to have checks while setting it.  I figure it is easy to filter it out so we never deal with it.