VFXACT-645 Adding support for creating connections between nodes.
Previously when you retrieved from scene the order of individual tissues was important because once a 'child' sub-tissue was added it would expect the parent to already be added.  This is fragile so this adds a method to explicitly add connections.  The intent is this runs after everything has been populated internally in zBuilder.  That removes requirement of individual node types coming in a specific order of that type.