VFXACT-539 Simplify method
Previously this was calling 'get_map_names' which just assembles the name.  Then with newly assembled name it gets_scene_item based on that name.

Now, the maps are stored per node in parameters dict.  So instead of assembling name and retrieving correct scene item, now we can loop through self.parameters['map'] which simplifies it as we already have correct map, no need to assemble name and retrieve again.