VFXACT-539 Removing of a couple methods
Removing get_mesh_objects as it was not used so removing.

Removing get_parameters as it has been simplified.  Calls using this use node.parameters instead.   

Removing get_map_objects.  This one was being used in a way that it is more efficient to simply use self.parameters as well.