VFXACT-539 Update to method to return a deepcopy instead of a shallow
return_copy_buffer() is meant for testing.  The idea is we can use this as soon as copy buffer is filled and then do some operations and use it to compare to copy buffer.  The copy buffer should not change.

Previously I was returning a shallow copy which always was true for a comparison so updated to return a deep copy.