VFXACT-768 Adding default return to find_class
When moving this from IO to builder, and refactoring node_factory in builder I forgot the case where the class object was not found.  It was returning None and causing problems.  Original behavior was to return a DG node object if no class is found to use that.  So now we are returning that if none are found.