VFXACT-637 Make ziva-vfx-utils Python version agnostic
The Maya 2022 support both Python 2 and Python 3,
the Python scripts in ziva-vfx-utils repo shall support both.
The changes are listed in ZivaVFX confluence pages,
"Python 2 -->Python 3 conversion".

Change some iterative methods, like xrange() to range(),
may affect the performance in Python 2 in some degree,
but that is also a good reason to switch to Python 3 earlier.