VFXACT-843 Add work around to get weight map function
In Maya 2022, it introduces a regression bug that
cause deformerSet Python API fail.
To work around this issue, add the fallback functions to the zBuilder.
The set/get_weights_fallback_impl() uses old getAttr() method.

The cost is the performance slow down when retrieve weight maps.
The unit test affected are modified to accomodate these changes.