Encapsulate weights fallback functions in paintable_maps.py
Localize the conditional logic based on Maya version in one place,
inside get_paintable_weights(). Now the callers of this function
don't need to check the Maya version before deciding whether to call
a fallback function, get_paintable_weights() does it for you. This
was already done in set_paintable_weights(), so now the two are
consistent.

The only wrinkle is that the signature of get_paintable_weights()
has expanded to include a mesh_name parameter which is needed by the
fallback. It's not hard to provide this in all the places we call
get_paintable_weights() in our code, but it is an API change to be
aware of.