VFXACT-964 Randomize license registration unit test temp file name
The license registration unit test uses some fixed file name.
This is fine is it runs alone.
But we add py2 and py3 unit test for Maya 2022 and run them parallel.
That causes race condition that make one process stomp the file written
by the other.
To prevent this problem, randomize the file names.