VFXACT-1004 Add create_subtree() helper function
One feature of creating Group node is create a new Group node
that contains selected nodes.
The create_subtree() is for supporting this.
It first checks all the input nodes,
prune those parent nodes that already in the node list.
Then create a new node as their parent node.

Unit test is added to verify the logic.
As Group node class has not implemented yet, a mockup class is added.