Fix bug of empty component treeview when select a zGeo tree item
This bug is caused by Python 2/3 difference on filter builtin.
It return list in Python 2 and generator in Python 3.
In the on_tvGeo_selected() callback, I use the same var name
to store the result of selected nodes and later use it
to store the filter generator that refer to the selected node list.
This overwrites the value that variable holds and thus return empty
list.

The bug only happens in Python 3 as Python 2 the filter() returns
a separate list, which is not affected.

Refer to this link for detail explanation,
https://stackoverflow.com/questions/41666977/python-2-vs-python-3-difference-in-behavior-of-filter