Make DGNode class hashable in Python 3
Refer to
https://docs.python.org/3/reference/datamodel.html#object.__hash__o
In Python 3 a class set its __hash__() to None,
if it defines __eq__() function.
The zBuilder.Base class is this case.
It makes derived DGNode class can't be used by set().

To fix this problem, define the Base.__hash__() function.
It doesn't implement a correct and robust hash function,
just return the id(self), to make set() keep working.