VFXACT-932 Refactor menubar code to be data driven
The menubar hierarchy now is data driven by a dict-tuple structure.
Each menu item is a key value pair, the key is the menu name,
and the value is the actions tuple, or a dict of submenu.
During the refactoring, I remove the "Cache" menu.
It is more suitable acts as a toolbar dropdown list button.