VFXACT-1092 Add load_plug_data flag to zGeoWidget.reset_builder()
This is the first step to support save/load scene panel data.
It adds a load_plug_data flag to zGeoWidget.reset_builder() function,
to decide whether to load the scene panel data before merge data.
There are some cases that need to set this flag:
1. When a Maya scene is opened and Scene Panel is notified by callback.
2. When Scene Panel is created.
   There maybe a Maya scene already opened and edited.
   The Scene Panel needs to load the data and merge it with latest
   zBuilder parse result.
The toolbar create function logic is updated to accommodate this change.
A new slot_argment is added to support slot function's arguments.