VFXACT-922 Group some component sections
Accodring to TD feedback, some tissue components need to combine together.
To support this, change the component_type_dict data structure,
it now supports tuple for component type.
The ComponentWidget.reset_model() and ComponentSectionWidget ctor
are updated to handle multiple component types.
The type name are concantenate to a list and their icon are added.

The icon button is used as fold the component section,
now I remove it because it's confusing when multiple icons are shown.
Instead, I will implement fold feature in other way.