VFXACT-922 Improve ComponentSection widget folding style
In last commit, the ComponentSection widget can fold/expand,
but the splitter handle does not attach to it when folded.

By appending an invisiable QFrame at the end,
and replace layout hint size height with actual control height,
the parent widget can compute each section height through
QSplitter.setSizes().

This is not perfect but make it look better.