VFXACT-1158 Fix "Internal C++ object already deleted." exception
After last fix, the error still occurs when press the Scene Panel button,
then load a demo scene.
The number of error occurs proportional to the number of launching Scene Panel.
The fix is to try catch this RuntimeError exception.

It's still a puzzle to me that even the Scene Panel is a singleton,
i.e., when a new Scene Panel is launched, the old one gets deleted and returned.
But the message callbacks seems not removed and still send messages.
That cause the reset_builder() gets called the number equals Scene Panel instantiated.