VFXACT-1059 Update Load/Save ziva rig doc
The Ziva rig load/save workflow are scripts in the ziva-vfx-utils repo,
the doc should be maintained in here too.
With the introduction of Scene Panel 2, these menu items and doc are copied.

This commits updates some sphinx grammar.
The skin cluster tranfer feature has not copied yet,
will be done in follow up commit.