VFXACT-1247 Refactor TreeItem.get_tree_path() to when non-root node as top node
The current TreeItem.get_tree_path() does not return correct tree path,
if the input top node is not root node.
For example, if the top node is zSolverTM, it returns "|",
not "|zSolver1", all the children nodes are affected too.
This causes serialize.construct_tree() fail to create tree item.

This commit refactors the function and fix tree path nameing issue.
A new unit test is added to verify this.