VFXACT-1268 Move builder.py to builders folder
The builder.py define the base class for all the derived builders.
It should be in the same folder as the others.
The bundle.py is the underlying data structure of the Builder class,
so it should reside the same place as the builder.py.
The IO.py is only used by builder.py but it also refer to
modules in the nodes folder.
It needs further refactoring to decouple the cross dependency.