VFXACT-1288 Refactor mesh.py
1. Remove verbose mesh class' accessor functions.
   Use the member variables directly in the member functions.
   Only the get_point_list() is kept because it is used by unit test.
   Will remove it in the future with get_intermediate().
2. Merge build_mesh() free function to its member function couterpart.
   Thanks to Maya Python API 2.0, the M*Array are the same as list type,
   which simplifies the logic.
3. Merge the get_mesh_vertex_positions() into get_mesh_info().

Lesson learnt:
Don't remove self.type assignment in the populate().
It looks like redundant but is necessary for the populating workflow.
Without it, the mesh class type is not initialized and cause problem.
This needs to be fixed in the future,
as each class has its own 'type' class variable.