VFXACT-1288 Fix zBuilder interplating map regression
When going through zBuilder tutorial, the feature of applying
interpolate map to smoothed mesh failed.
The cause is commit 989d00c, which we tried to get Maya's intermediate
mesh for non-start frame mesh result.
Revert that change for now as that is not a good solution.
We will think of other solution for that task, which is logged as VFXACT-356.

A unit test is added to prevent regression.