VFXACT-1813 Revise zBuilder.commands docstring and documentation
One of our internal user spotted a name convention issue in the zBuilder doc.
Further check shows that the doc texts are copied from
function docstring defined in the zBuilder.command module.
To avoid maintain 2 copies of the texts, I replace the example code
and comments in the documentation with the cross reference to the
functions in the zBuilder.commands module.
By doing that, we only need to maintain the docstring in the future.

Another important change is bump the zBuilder version number in the doc,
which is forgot during v2.2 release.