VFXACT-1875 Fix sprase array plug value cause length mismatch with mesh vertices length
When calling get_paintable_map() with Blendshape,
its return weight values length doesn't match the mesh vertcies length.
The cause is the array plug that store the weight values is sparse.
It doesn't store the default values.
When read the plug values, it only returns plug value that is different
than the default one.

To fix this issue, get_paintable_map_by_ArrayDataBuilder() add mesh name
as new parameter to retrieve the mesh vertex count.
The return weight list is not empty by default, instead,
it is initialized with the default value that is queried from array plug,
filled with mesh vertex count.
When reading the array plug, set the value on the logic index get from the plug.

Unit test is added to cover the new behavior.
The bonewarp test case is updated as well, the mesh name is a needed
parameter that is used to get the mesh vertex count.
We can't get rid of it even we fix the MFnWeightGeometryFilter regression.

The simple performance test is to create Blendshape with polycube that is
subdivided to 100 on xyz direction, paint some weight values.
The get_paintable_map_by_ArrayDataBuilder() takes 0.094s,
while the _get_paintable_map_by_MFnWeightGeometryFilter_fallback_impl()
takes 23.953s.

The faster version is still better than the fallback method.